get interrupts and scheduler working
