+++ include/irq.h
@@ -3,7 +3,7 @@
 
 void enable_interrupt_controller( void );
 
-extern "C" void irq_vector_init( void );
+extern "C" void handle_irq( void );
 extern "C" void enable_irq( void );
 extern "C" void disable_irq( void );
