+++ include/sched.h
@@ -9,9 +9,6 @@
 
 #define NR_TASKS				64 
 
-#define FIRST_TASK task[0]
-#define LAST_TASK task[NR_TASKS-1]
-
 #define TASK_RUNNING				0
 
 extern struct task_struct *current;
@@ -51,8 +48,8 @@ extern "C" void switch_to(struct task_struct* next);
 extern "C" void cpu_switch_to(struct task_struct* prev, struct task_struct* next);
 
 #define INIT_TASK \
-/*cpu_context*/	{ {0,0,0,0,0,0,0,0,0,0,0,0,0}, \
-/* state etc */	0,0,1, 0 \
+{ {0,0,0,0,0,0,0,0,0,0,0,0,0}, \
+0,0,1,0 \
 }
 
 #endif
