+++ src/boot.S
@@ -115,6 +115,47 @@ setup_el1_for_secondary:
     // Transition to EL1
     eret
 
+.macro	kernel_entry
+	sub	sp, sp, #S_FRAME_SIZE
+	stp	x0, x1, [sp, #16 * 0]
+	stp	x2, x3, [sp, #16 * 1]
+	stp	x4, x5, [sp, #16 * 2]
+	stp	x6, x7, [sp, #16 * 3]
+	stp	x8, x9, [sp, #16 * 4]
+	stp	x10, x11, [sp, #16 * 5]
+	stp	x12, x13, [sp, #16 * 6]
+	stp	x14, x15, [sp, #16 * 7]
+	stp	x16, x17, [sp, #16 * 8]
+	stp	x18, x19, [sp, #16 * 9]
+	stp	x20, x21, [sp, #16 * 10]
+	stp	x22, x23, [sp, #16 * 11]
+	stp	x24, x25, [sp, #16 * 12]
+	stp	x26, x27, [sp, #16 * 13]
+	stp	x28, x29, [sp, #16 * 14]
+	str	x30, [sp, #16 * 15] 
+	.endm
+
+	.macro	kernel_exit
+	ldp	x0, x1, [sp, #16 * 0]
+	ldp	x2, x3, [sp, #16 * 1]
+	ldp	x4, x5, [sp, #16 * 2]
+	ldp	x6, x7, [sp, #16 * 3]
+	ldp	x8, x9, [sp, #16 * 4]
+	ldp	x10, x11, [sp, #16 * 5]
+	ldp	x12, x13, [sp, #16 * 6]
+	ldp	x14, x15, [sp, #16 * 7]
+	ldp	x16, x17, [sp, #16 * 8]
+	ldp	x18, x19, [sp, #16 * 9]
+	ldp	x20, x21, [sp, #16 * 10]
+	ldp	x22, x23, [sp, #16 * 11]
+	ldp	x24, x25, [sp, #16 * 12]
+	ldp	x26, x27, [sp, #16 * 13]
+	ldp	x28, x29, [sp, #16 * 14]
+	ldr	x30, [sp, #16 * 15] 
+	add	sp, sp, #S_FRAME_SIZE		
+	eret
+	.endm
+
 secondary_kernel_main:
     bl kernel_init
     b proc_hang
@@ -132,41 +173,116 @@ wake_up_cores:
     sev
     ret
 
+
+
+.macro ventry label
+.align	7
+b	\label
+.endm
+
+
     // important, code has to be properly aligned
     .align 11
 _vectors:
-    // synchronous
-    .align  7
+    ventry synchronous_el0
+    ventry irq_el0
+    ventry fiq_el0
+    ventry serror_el0
+    
+    ventry synchronous_el1
+    ventry irq_el1
+    ventry fiq_el1
+    ventry serror_el1
+    
+    ventry synchronous_el0_64
+    ventry irq_el0_64
+    ventry fiq_el0_64
+    ventry serror_el0_64
+    
+    ventry synchronous_el0_32
+    ventry irq_el0_32
+    ventry fiq_el0_32
+    ventry serror_el0_32
+
+.macro handle_exception
+mrs     x1, esr_el1
+mrs     x2, elr_el1
+mrs     x3, spsr_el1
+mrs     x4, far_el1
+b       exc_handler
+.endm
+
+synchronous_el0:
     mov     x0, #0
-    mrs     x1, esr_el1
-    mrs     x2, elr_el1
-    mrs     x3, spsr_el1
-    mrs     x4, far_el1
-    b       exc_handler
+    handle_exception
 
-    // IRQ
-    .align  7
-    mov     x0, #1
-    mrs     x1, esr_el1
-    mrs     x2, elr_el1
-    mrs     x3, spsr_el1
-    mrs     x4, far_el1
-    b       exc_handler
+irq_el0: 
+	kernel_entry 
+	bl	handle_irq
+	kernel_exit 
 
-    // FIQ
-    .align  7
+fiq_el0:
     mov     x0, #2
-    mrs     x1, esr_el1
-    mrs     x2, elr_el1
-    mrs     x3, spsr_el1
-    mrs     x4, far_el1
-    b       exc_handler
+    handle_exception
 
-    // SError
-    .align  7
+serror_el0:
     mov     x0, #3
-    mrs     x1, esr_el1
-    mrs     x2, elr_el1
-    mrs     x3, spsr_el1
-    mrs     x4, far_el1
-    b       exc_handler
+    handle_exception
+    
+
+synchronous_el1:
+    mov     x0, #0
+    handle_exception
+
+irq_el1: 
+	kernel_entry 
+	bl	handle_irq
+	kernel_exit 
+
+fiq_el1:
+    mov     x0, #2
+    handle_exception
+
+serror_el1:
+    mov     x0, #3
+    handle_exception
+    
+synchronous_el0_64:
+    mov     x0, #0
+    handle_exception
+
+irq_el0_64: 
+	kernel_entry 
+	bl	handle_irq
+	kernel_exit 
+
+fiq_el0_64:
+    mov     x0, #2
+    handle_exception
+
+serror_el0_64:
+    mov     x0, #3
+    handle_exception
+    
+synchronous_el0_32:
+    mov     x0, #0
+    handle_exception
+
+irq_el0_32: 
+	kernel_entry 
+	bl	handle_irq
+	kernel_exit 
+
+fiq_el0_32:
+    mov     x0, #2
+    handle_exception
+
+serror_el0_32:
+    mov     x0, #3
+    handle_exception
+
+.globl ret_from_fork
+ret_from_fork:
+	bl	schedule_tail
+	mov	x0, x20
+	blr	x19 		//should never return
