+++ src/sched.cpp
@@ -7,19 +7,15 @@ struct task_struct *current = &(init_task);
 struct task_struct * task[NR_TASKS] = {&(init_task), };
 int nr_tasks = 1;
 
-void preempt_disable(void)
-{
+void preempt_disable(void) {
 	current->preempt_count++;
 }
 
-void preempt_enable(void)
-{
+void preempt_enable(void) {
 	current->preempt_count--;
 }
 
-
-void _schedule(void)
-{
+void _schedule(void) {
 	preempt_disable();
 	int next,c;
 	struct task_struct * p;
@@ -27,7 +23,6 @@ void _schedule(void)
 		c = -1;
 		next = 0;
 		for (int i = 0; i < NR_TASKS; i++){
-			printf("task %d is %d\n", i, task[i]);
 			p = task[i];
 			if (p && p->state == TASK_RUNNING && p->counter > c) {
 				c = p->counter;
@@ -44,19 +39,16 @@ void _schedule(void)
 			}
 		}
 	}
-	printf("WE ARE GONNA SWITCH LFG to %d\n", task[next]);
 	switch_to(task[next]);
 	preempt_enable();
 }
 
-void schedule(void)
-{
+void schedule(void) {
 	current->counter = 0;
 	_schedule();
 }
 
-extern "C" void switch_to(struct task_struct * next) 
-{
+extern "C" void switch_to(struct task_struct * next) {
 	if (current == next) 
 		return;
 	struct task_struct * prev = current;
@@ -69,13 +61,12 @@ extern "C" void schedule_tail(void) {
 }
 
 
-extern "C" void timer_tick()
-{
+extern "C" void timer_tick() {
 	--current->counter;
-	if (current->counter>0 || current->preempt_count >0) {
+	if (current->counter > 0 || current->preempt_count > 0) {
 		return;
 	}
-	current->counter=0;
+	current->counter = 0;
 	enable_irq();
 	_schedule();
 	disable_irq();
